1.2.0 RELEASE FINAL

+++ Model/Listing.php
@@ -447,7 +447,9 @@ class Listing extends \Ess\M2ePro\Model\ActiveRecord\Component\Parent\AbstractMo
 
     //########################################
 
-    public function addProduct($product, $checkingMode = false, $checkHasProduct = true,
+    public function addProduct($product,
+                               $initiator = \Ess\M2ePro\Helper\Data::INITIATOR_UNKNOWN,
+                               $checkingMode = false, $checkHasProduct = true,
                                array $logAdditionalInfo = array())
     {
         $productId = $product instanceof \Magento\Catalog\Model\Product ?
@@ -485,7 +487,7 @@ class Listing extends \Ess\M2ePro\Model\ActiveRecord\Component\Parent\AbstractMo
             $this->getId(),
             $productId,
             $listingProductTemp->getId(),
-            \Ess\M2ePro\Helper\Data::INITIATOR_UNKNOWN,
+            $initiator,
             NULL,
             \Ess\M2ePro\Model\Listing\Log::ACTION_ADD_PRODUCT_TO_LISTING,
             // M2ePro\TRANSLATIONS
@@ -502,11 +504,12 @@ class Listing extends \Ess\M2ePro\Model\ActiveRecord\Component\Parent\AbstractMo
 
     // ---------------------------------------
 
-    public function addProductsFromCategory($categoryId)
+    public function addProductsFromCategory($categoryId,
+                                            $initiator = \Ess\M2ePro\Helper\Data::INITIATOR_UNKNOWN)
     {
         $categoryProductsArray = $this->getProductsFromCategory($categoryId);
         foreach ($categoryProductsArray as $productTemp) {
-            $this->addProduct($productTemp);
+            $this->addProduct($productTemp, $initiator);
         }
     }
 
@@ -609,17 +612,73 @@ class Listing extends \Ess\M2ePro\Model\ActiveRecord\Component\Parent\AbstractMo
                                     ->addFieldToFilter('product_id', $productId)
                                     ->getItems();
 
-        $deletedVariationsIds = array();
+        $processedVariationsIds = array();
 
         /** @var $variationOption \Ess\M2ePro\Model\Listing\Product\Variation\Option */
         foreach ($variationOptions as $variationOption) {
 
+            if (in_array($variationOption->getListingProductVariationId(), $processedVariationsIds)) {
+                continue;
+            }
+
+            $processedVariationsIds[] = $variationOption->getListingProductVariationId();
+
             /** @var $listingProduct \Ess\M2ePro\Model\Listing\Product */
             $listingProduct = $variationOption->getListingProduct();
 
-            if (!in_array($variationOption->getListingProductVariationId(),$deletedVariationsIds)) {
-                $variationOption->getListingProductVariation()->delete();
-                $deletedVariationsIds[] = $variationOption->getListingProductVariationId();
+            if ($variationOption->isComponentModeEbay()) {
+
+                /** @var \Ess\M2ePro\Model\Ebay\Listing\Product\Variation $ebayVariation */
+                $variation = $variationOption->getListingProductVariation();
+                $ebayVariation = $variation->getChildObject();
+
+                if (!$ebayVariation->isNotListed()) {
+                    $additionalData = $listingProduct->getAdditionalData();
+                    $variationsThatCanNotBeDeleted = isset($additionalData['variations_that_can_not_be_deleted'])
+                        ? $additionalData['variations_that_can_not_be_deleted'] : array();
+
+                    $specifics = array();
+
+                    foreach ($variation->getOptions(true) as $option) {
+                        $specifics[$option->getAttribute()] = $option->getOption();
+                    }
+
+                    $tempVariation[] = array(
+                        'qty' => 0,
+                        'price' => $ebayVariation->getPrice(),
+                        'sku' => $ebayVariation->getOnlineSku(),
+                        'add' => 0,
+                        'delete' => 1,
+                        'specifics' => $specifics,
+                        'has_sales' => $ebayVariation->hasSales()
+                    );
+
+                    $key = 'variations_specifics_replacements';
+                    if (!empty($additionalData[$key])) {
+                        $tempVariation[$key] = $additionalData[$key];
+                    }
+
+                    $variationAdditionalData = $variation->getAdditionalData();
+                    if (isset($variationAdditionalData['ebay_mpn_value'])) {
+                        $tempVariation['details']['mpn'] = $variationAdditionalData['ebay_mpn_value'];
+                    }
+
+                    $variationsThatCanNotBeDeleted[] = $tempVariation;
+                    $additionalData['variations_that_can_not_be_deleted'] = $variationsThatCanNotBeDeleted;
+
+                    $listingProduct->setSettings('additional_data', $additionalData)->save();
+                }
+
+                $variation->delete();
+            } else {
+                $listingProduct->deleteProcessingLocks();
+
+                if ($listingProduct->isStoppable()) {
+                    $this->activeRecordFactory->getObject('StopQueue')->add($listingProduct);
+                    $listingProduct->setStatus(\Ess\M2ePro\Model\Listing\Product::STATUS_STOPPED)->save();
+                }
+
+                $listingsProductsForRemove[$listingProduct->getId()] = $listingProduct;
             }
 
             $listingId = $listingProduct->getListingId();
@@ -647,7 +706,24 @@ class Listing extends \Ess\M2ePro\Model\ActiveRecord\Component\Parent\AbstractMo
         }
 
         foreach ($listingsProductsForRemove as $listingProduct) {
-            $listingProduct->delete();
+
+            if ($listingProduct->isComponentModeAmazon()) {
+                /** @var \Ess\M2ePro\Model\Amazon\Listing\Product $amazonListingProduct */
+                $amazonListingProduct = $listingProduct->getChildObject();
+                $variationManager = $amazonListingProduct->getVariationManager();
+
+                if ($variationManager->isRelationChildType()) {
+                    /** @var \Ess\M2ePro\Model\Amazon\Listing\Product $amazonParentListingProduct */
+                    $amazonParentListingProduct = $variationManager->getTypeModel()->getAmazonParentListingProduct();
+
+                    $listingProduct->delete();
+
+                    $amazonParentListingProduct->getVariationManager()->getTypeModel()->getProcessor()->process();
+                }
+            } else {
+                $listingProduct->delete();
+            }
+
         }
         // ---------------------------------------
     }
